lab_02 docs + fixes

+++ lab_01/Encoder.h
@@ -10,14 +10,14 @@ class Encoder
 {
 
 public:
-    Encoder(uint8_t alphabet_size, std::string new_alphabet);
-    Encoder(uint8_t alphabet_size, std::vector<uint8_t> new_alphabet);
+    Encoder(uint16_t alphabet_size, std::string new_alphabet);
+    Encoder(uint16_t alphabet_size, std::vector<uint8_t> new_alphabet);
 
     uint8_t encode(uint8_t symbol);
     uint8_t decode(uint8_t code);
 
 private:
-    uint8_t size_alpha;
+    uint16_t size_alpha;
     std::vector<uint8_t> alphabet;
 };
