+++ lab_01/Enigma.h
@@ -9,8 +9,8 @@ class Enigma
 {
 
 public:
-    Enigma(uint8_t size_rotor,
-           uint8_t num_rotors,
+    Enigma(uint16_t size_rotor,
+           uint16_t num_rotors,
            const Reflector& reflector_place,
            const std::vector<Rotor>& rotors_place,
            const Commutator& commutator_place);
@@ -24,8 +24,8 @@ private:
 
 private:
     uint64_t counter = 0;
-    uint8_t size_rotor;
-    uint8_t num_rotors;
+    uint16_t size_rotor;
+    uint16_t num_rotors;
     Reflector reflector;
     Commutator commutator;
     std::vector<Rotor> rotors;
